+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/FoundryProperties.kt
@@ -405,10 +405,6 @@ internal constructor(
   public val enableMoshiIr: Boolean
     get() = booleanProperty("foundry.moshi.ir.enable")
 
-  /** Flag to enable/disable moshi proguard rule gen. */
-  public val moshixGenerateProguardRules: Boolean
-    get() = booleanProperty("moshix.generateProguardRules", defaultValue = true)
-
   /** Flag to connect ViewBinding sources to KSP. */
   public val kspConnectViewBinding: Boolean
     get() = booleanProperty("foundry.ksp.connect.viewbinding")
@@ -906,9 +902,6 @@ internal constructor(
 
   public val compileSdkVersion: String?
     get() = optionalStringProperty("foundry.android.compileSdkVersion")
-
-  public fun latestCompileSdkWithSources(defaultValue: Int): Int =
-    intProperty("foundry.android.latestCompileSdkWithSources", defaultValue = defaultValue)
 
   private val minSdkVersion: String?
     get() = optionalStringProperty("foundry.android.minSdkVersion")