+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/topography/ModuleTopographyTask.kt
@@ -111,7 +111,9 @@ internal object ModuleTopographyTasks {
           },
         )
         features.put(
-          DefaultFeatures.DaggerCompiler,
+          (foundryProperties.anvilMode.useDaggerKsp.let {
+            if (it) DefaultFeatures.DaggerCompilerKsp else DefaultFeatures.DaggerCompilerKapt
+          }),
           foundryExtension.featuresHandler.diHandler.useDaggerCompiler,
         )
         features.put(