+++ platforms/gradle/foundry-gradle-plugin/src/main/kotlin/foundry/gradle/util/kgpUtil.kt
@@ -15,16 +15,14 @@
  */
 package foundry.gradle.util
 
-import com.google.devtools.ksp.gradle.KspTask
 import org.gradle.api.tasks.TaskContainer
-import org.jetbrains.kotlin.gradle.internal.KaptGenerateStubsTask
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompilationTask
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 /**
- * Configures [KotlinCompile] tasks with the given [action] but _ignores_ [KaptGenerateStubsTask]
- * and [KspTask] types as they inherit arguments from standard compilation tasks via and applying
- * arguments to them could result in duplicates.
+ * Configures [KotlinCompile] tasks with the given [action] but _ignores_ KaptGenerateStubsTask and
+ * KSP types as they inherit arguments from standard compilation tasks via and applying arguments to
+ * them could result in duplicates.
  *
  * See
  * https://github.com/JetBrains/kotlin/blob/0e4e53786c1b0341befe8e71a5e6e0bc0e464370/libraries/tools/kotlin-gradle-plugin/src/common/kotlin/org/jetbrains/kotlin/gradle/internal/kapt/KaptGenerateStubsTask.kt#L111-L113
@@ -38,8 +36,8 @@ internal fun TaskContainer.configureKotlinCompilationTask(
     // Kapt stub gen is a special case because KGP sets it up to copy compiler args from the
     // standard kotlin compilation, which can lead to duplicates. SOOOO we skip configuration of
     // it here. Callers to this _can_ opt in to including it, but they must be explicit.
-    if (includeKaptGenerateStubsTask || this !is KaptGenerateStubsTask) {
-      if (includeKspTask || this !is KspTask) {
+    if (includeKaptGenerateStubsTask || !this.name.startsWith("kapt")) {
+      if (includeKspTask || !this.name.startsWith("ksp")) {
         action()
       }
     }